create donation post details card and make reusableWidgets directory
